requirements.txt
snowflake-connector-python==3.12.2
anthropic==0.39.0
crewai==0.86.0
crewai-tools==0.17.0
streamlit==1.39.0
python-dotenv==1.0.1
pandas==2.2.3